+++ test/cases/css-devtool-source-map/webpack.config.js
@@ -19,7 +19,19 @@ module.exports = {
     rules: [
       {
         test: /\.(css|scss)$/,
-        use: ['css-loader', 'sass-loader'],
+        //use: ['css-loader', 'sass-loader'], // sass-loader <= 15.0.0
+        use: [
+          'css-loader',
+          {
+            loader: 'sass-loader',
+            options: {
+              sassOptions: {
+                // include the sources in the generated source map (required since sass-loader >= 16.0.0)
+                sourceMapIncludeSources: true,
+              },
+            },
+          },
+        ],
       },
     ],
   },