+++ test/cases/entry-scss-font-url/webpack.config.js
@@ -25,7 +25,19 @@ module.exports = {
     rules: [
       {
         test: /\.(css|sass|scss)$/,
-        use: ['css-loader', 'sass-loader'],
+        //use: ['css-loader', 'sass-loader'], // sass-loader <= 15.0.0
+        use: [
+          'css-loader',
+          {
+            loader: 'sass-loader',
+            options: {
+              sassOptions: {
+                // include the sources in the generated source map (required since sass-loader >= 16.0.0)
+                sourceMapIncludeSources: true,
+              },
+            },
+          },
+        ],
       },
       {
         test: /\.(woff|woff2|eot|ttf|otf|svg)$/i,