+++ test/cases/split-chunk-resolve-assets-prod/expected/assets/js/552.a12a7ed5.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[171,552],{171:(o,l,e)=>{const{lorem:n,libA:s,libB:c}=e(854),g=e(198),a=e(740);console.log(">> main-a:"),console.log(" - A: ",g),console.log(" - C: ",a),console.log("Lorem: ",n.getTitle())}}]);